+++ CHANGELOG.md
@@ -17,6 +17,7 @@ New:
 * Add PeriodicTransformXD and VectorPeriodicTransformXD functions to support the data simulated with periodic boundary conditions. (#387)
 * Add CylindricalTransform and VectorCylindricalTransform to transform functions from cylindrical to Cartesian coordinates. (#387)
 * StarkBroadenedLine now supports Doppler broadening and Zeeman splitting. (#393)
+* Add the kind attribute to RayTransferPipelineXD that determines whether the ray transfer matrix is multiplied by sensitivity ('power') or not ('radiance'). (#412)
 
 
 Release 1.4.0 (3 Feb 2023)